# src/swagger-config.yaml
# https://swagger.io/docs/specification/2-0
swagger: "2.0"
info:
  version: "1.0.0"
  title: "Swagger GeoPick docs"
  description: "List of all API endpoints"
host: 127.0.0.1:8085 
basePath: /v1
schemes:
  - http
paths:
  /authenticate:
    post:
      tags: 
        - "Authentication"
      summary: "Provides a token valid for use against authenticated endpoints"
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: "access data"
          in: body
          description: "JSON data for the api"
          required: true
          schema:
            type: object
            properties: 
              username: 
                type: string          
                description: "Valid username of registered user"
              password:
                type: string
                description: "Valid password of registered user"

      responses:
        404:
          description: "No user with these credentials exist"
        200:
          description: "Authentication successful"
  /sec:
    post:
      tags: 
        - "Web application private methods"
      summary: "This method is intended to be used only by the GeoPick web application. To directly query the API please refer to the 'georeference-dwc' endpoint."
      description: "Given an incoming spatial geometry in GeoJSON format returns its complete point-radius georeference including its SEC, in a format to be consumed by the GeoPick application. This is a method intended to be used only by the GeoPick web application."
      consumes:
        - application/json  # Specify the request content type
      produces:
        - application/json
      parameters:
        - name: "access data"
          in: body
          description: "JSON data for the api"
          required: true
          schema:
            type: object
            properties:
              type: 
                type: string
              geometry:
                type: object
                properties:
                  type:
                    type: string  # "type" is a string
                  coordinates:
                    type: array
                    items:
                      type: array
                      items:
                        type: number  # Coordinates are represented as numbers

      responses:
        200:
          description: "OK"
        400:
          description: "Bad Request"  

  /georeference-dwc:
    post:
      tags: 
        - "API ßmethods"
      summary: "Given a georeferenced locality expressed in GeoJSON format, returns a complete point-radius georeference as a GeoJSON object whose keys are named in Darwin Core named terms. It also includes a polygonal representation of the smallest enclosing circle."
      consumes:
        - application/json  # Specify the request content type
      produces:
        - application/json
      parameters:
        - name: "access data"
          in: body
          description: "JSON data for the api"
          required: true
          schema:
            type: object
            properties:
              locality: 
                type: string
              type:
                type: string
              georeferencedBy:
                type: string
              georeferenceRemarks:
                type: string      
              features:
                type: object
                properties:
                  id:
                    type: string  # "id" is a string in this case
                  type:
                    type: string  # "type" is assumed to be a string
                  properties:
                    type: object  # "properties" is an object
                  geometry:
                    type: object
                    properties:
                      type:
                        type: string  # "type" is a string
                      coordinates:
                        type: array
                        items:
                          type: array
                          items:
                            type: number  # Coordinates are represented as numbers
            
      responses:
        200:
          description: "OK"
        400:
          description: "Bad Request"  